tidyfeed-python-worker/requirements.txt
requests>=2.31.0
yt-dlp>=2024.1.0
boto3>=1.34.0
python-dotenv>=1.0.0
twikit>=2.0.0
